Edicion de setup.php

+++ setup.php
@@ -2,16 +2,23 @@
 // BORRAR DESPUÉS DE USAR
 $composerPhar = __DIR__ . '/composer.phar';
 
-// Descargar composer.phar
-file_put_contents(
-    $composerPhar,
-    file_get_contents('https://getcomposer.org/composer-stable.phar')
-);
+if (!file_exists($composerPhar)) {
+    file_put_contents(
+        $composerPhar,
+        file_get_contents('https://getcomposer.org/composer-stable.phar')
+    );
+}
 
-// Ejecutar composer dump-autoload
+putenv('HOME=' . __DIR__);
+putenv('COMPOSER_HOME=' . __DIR__ . '/.composer');
+
 $output = [];
 $return = 0;
-exec('php ' . $composerPhar . ' dump-autoload --optimize --working-dir=' . __DIR__ . ' 2>&1', $output, $return);
+exec(
+    'HOME=' . __DIR__ . ' COMPOSER_HOME=' . __DIR__ . '/.composer php ' . $composerPhar . ' dump-autoload --optimize --working-dir=' . __DIR__ . ' 2>&1',
+    $output,
+    $return
+);
 
 echo '<pre>';
 echo 'Return code: ' . $return . "\n";
